+++ clients/sirene-insee/export-sirene.ts
@@ -1,7 +1,7 @@
 import routes from '#clients/routes';
 import { exportCsvClientGet } from '#clients/sirene-insee';
 import constants from '#models/constants';
-import { ExportCsvInput } from 'app/api/export-csv/input-validation';
+import { ExportCsvInput } from 'app/api/export-sirene/input-validation';
 import { Readable } from 'stream';
 import { SireneQueryBuilder } from './build-query';
